+++ Resources/Prototypes/Entities/Structures/Machines/lathe.yml
@@ -851,6 +851,7 @@
       - Saw
       - Hemostat
       - ClothingEyesGlassesChemical
+      - MedAutoinjector #ss220 Autoinjecors update
     dynamicRecipes:
       - ChemicalPayload
       - CryostasisBeaker
