+++ Resources/Prototypes/Recipes/Lathes/medical.yml
@@ -227,10 +227,11 @@
     Plastic: 300
 
 - type: latheRecipe
-  id: MedAutoinjectorr
+  id: MedAutoinjector
   result: MedAutoinjector
   completetime: 2
   materials:
-    Glass: 200
+    Glass: 100
     Plastic: 200
+    Steel: 25
